move the onionproxy package to within the service package
